on the game sumary tab, start new game should really say go back to title screen. on the title screen lets implment the ability to save and select up to 4 games. it should put a score and the round info in the sumary to help users know what game it is. dont let the user keep more than 4 saved games. but give them a delete (trash can) button so they can delete a game save. if the user trys to start a new game while there is 4 saves already. dont let them but give them a toast message to indicate they need to delete or finish one of their saved games.
lets build out the idea of finishing a game too. this should be triggerable from the game sumary tab as well with a quit game button. lets prompt the user to make sure they are okay with quiting and not saving the progress.

when the game ends the quit button should disapear. the workflow should be the user has to go back to the title screen. and when they do the finished game is cleared from memory and one of the 4 slots